IA-2468: move ExternalTaskSerializer out of polio
